requirements.txt
flask==2.0.1
torch==2.0.0
diffusers==0.11.1
open3d==0.17.0
numpy==1.21.0
Pillow==9.0.0
opencv-python==4.5.4.60
point-e==0.0.0
werkzeug==2.0.1
transformers>=4.30.2
python-dotenv>=0.19.0
requests>=2.31.0 